refactor: add API config / localStorage util func
