+++ src/config/api.js
@@ -1,0 +1,6 @@
+const BASE_URL = 'http://54.180.98.209:8000';
+
+export const API = {
+  SIGNUP: `${BASE_URL}/accounts/user/signup/`,
+  SIGNIN: `${BASE_URL}/accounts/user/signin/`,
+};
