--- src/constants/url.js
@@ -1,1 +1,0 @@
-export const baseURL = 'http://54.180.98.209:8000';
