+++ src/hooks/queries/useUser.jsx
@@ -1,13 +1,12 @@
 import { useQuery } from '@tanstack/react-query';
 import { useEffect } from 'react';
-import * as userLocalStorage from './userLocalstore';
+import * as userLocalStorage from '../../utils/userLocalstore';
 
 // 백한테 user 정보 달라고 요청해야함 추후에 회의 후 수정
 function getUser(user) {
   if (!user) return null;
   //   const { data } = await axiosInstance.get(`/user/${user.id}`);
   //   return data.user;
-  console.log(user);
   return {};
 }
